:wrench: Update launcher script

+++ run.sh
@@ -1,9 +1,20 @@
 #!/bin/sh
 
+[ ! $# -eq 0 ] || exit
+target=$1
+
+[ ! -e $target ] && make $target
+[ -e $target ] || exit
+
+xephyr=$(ps -ef | grep Xephyr | grep -v "grep" | wc -l)
+
+if [ $xephyr -eq 0 ]
+then
+    Xephyr -br -ac -reset -screen 1280x720 :1 &
+    sleep 1s
+
+    export DISPLAY=:1
+fi
+
+./$target
 pkill -f Xephyr
-Xephyr -br -ac -reset -screen 1280x720 :1 &
-
-sleep 1s
-export DISPLAY=:1
-
-./$1
